neem-interface.pl: Don't assert both is_action and has_type(...)
When creating an action and specifying its type with has_type,
is_action is redundant because the fact that it is an action is
already implicit in the type. Asserting both leads to the action
individual being an individual of BOTH dul:Action and the custom action
type, which would make subsequent calls to is_individual return 2
solutions. That is not wrong, but makes querying action types
unnecessarily complex.